Move test to common catalog
